+++ apps/web/jest.config.ts
@@ -2,7 +2,8 @@ import config from "@umami/jest-config";
 
 export default {
   ...config,
-  setupFilesAfterEnv: ["<rootDir>/src/setupTests.ts"],
+  setupFiles: ["<rootDir>/src/setupTests.ts"],
+  setupFilesAfterEnv: ["<rootDir>/src/setupTestsAfterEnv.ts"],
   rootDir: "./",
   bail: false, // TODO: remove
   transformIgnorePatterns: ["node_modules/(?!(redux-persist-transform-encrypt)/)"],
